refactor(agent): 通过 __init__.py 导出公开接口，迁移所有调用点
agent/__init__.py 导出 build_agent、Sonetto、build_system_prompt、
get_system_prompt_parts。6 处调用点统一改为 from agent import ...，
不再直接引用子模块。
